improve overlay handling
